Fix import cycle lint errors Enable linting on utils package
