+++ .eslintrc
@@ -14,6 +14,8 @@
   "plugins": ["@typescript-eslint","jest"],
   "extends": [
     "eslint:recommended",
+    "plugin:import/recommended", 
+    "plugin:import/typescript",
     "plugin:@typescript-eslint/recommended",
     "prettier"
   ],
@@ -63,5 +65,13 @@
     "jest/no-identical-title": "error",
     "jest/prefer-to-have-length": "warn",
     "jest/valid-expect": "error"
+  },
+  "settings": {
+    "import/resolver": {
+      // You will also need to install and configure the TypeScript resolver
+      // See also https://github.com/import-js/eslint-import-resolver-typescript#configuration
+      "typescript": true,
+      "node": true
+    }
   }
 }
