+++ package.json
@@ -8,6 +8,8 @@
     "@typescript-eslint/parser": "^7.4.0",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
+    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jest": "^28.2.0",
     "husky": "^8.0.0",
     "lint-staged": "^12.4.3",
