+++ typescript/sdk/src/ism/metadata/builder.ts
@@ -1,53 +1,22 @@
 /* eslint-disable no-case-declarations */
-import { TransactionReceipt } from '@ethersproject/providers';
-
 import { WithAddress, assert, rootLogger } from '@hyperlane-xyz/utils';
 
 import { deepFind } from '../../../../utils/dist/objects.js';
 import { HyperlaneCore } from '../../core/HyperlaneCore.js';
-import { DispatchedMessage } from '../../core/types.js';
-import { DerivedHookConfig } from '../../hook/EvmHookReader.js';
 import {
   ArbL2ToL1HookConfig,
   HookType,
   MerkleTreeHookConfig,
 } from '../../hook/types.js';
 import { MultiProvider } from '../../providers/MultiProvider.js';
-import { DerivedIsmConfig } from '../EvmIsmReader.js';
 import { IsmType } from '../types.js';
 
-import {
-  AggregationMetadata,
-  AggregationMetadataBuilder,
-} from './aggregation.js';
-import { ArbL2ToL1Metadata, ArbL2ToL1MetadataBuilder } from './arbL2ToL1.js';
-import { MultisigMetadata, MultisigMetadataBuilder } from './multisig.js';
-import { NullMetadata, NullMetadataBuilder } from './null.js';
-import {
-  DefaultFallbackRoutingMetadataBuilder,
-  RoutingMetadata,
-} from './routing.js';
-
-export type StructuredMetadata =
-  | NullMetadata
-  | MultisigMetadata
-  | ArbL2ToL1Metadata
-  | AggregationMetadata<any>
-  | RoutingMetadata<any>;
-
-export interface MetadataContext<
-  IsmContext = DerivedIsmConfig,
-  HookContext = DerivedHookConfig,
-> {
-  message: DispatchedMessage;
-  dispatchTx: TransactionReceipt;
-  ism: IsmContext;
-  hook: HookContext;
-}
-
-export interface MetadataBuilder {
-  build(context: MetadataContext): Promise<string>;
-}
+import { AggregationMetadataBuilder } from './aggregation.js';
+import { ArbL2ToL1MetadataBuilder } from './arbL2ToL1.js';
+import { MultisigMetadataBuilder } from './multisig.js';
+import { NullMetadataBuilder } from './null.js';
+import { DefaultFallbackRoutingMetadataBuilder } from './routing.js';
+import type { MetadataBuilder, MetadataContext } from './types.js';
 
 export class BaseMetadataBuilder implements MetadataBuilder {
   public nullMetadataBuilder: NullMetadataBuilder;
@@ -130,39 +99,6 @@ export class BaseMetadataBuilder implements MetadataBuilder {
 
       default:
         throw new Error(`Unsupported ISM: ${ism}`);
-    }
-  }
-
-  static decode(
-    metadata: string,
-    context: MetadataContext,
-  ): StructuredMetadata {
-    const { ism } = context;
-    switch (ism.type) {
-      case IsmType.TRUSTED_RELAYER:
-        return NullMetadataBuilder.decode(ism);
-
-      case IsmType.MERKLE_ROOT_MULTISIG:
-      case IsmType.MESSAGE_ID_MULTISIG:
-        return MultisigMetadataBuilder.decode(metadata, ism.type);
-
-      case IsmType.AGGREGATION:
-        return AggregationMetadataBuilder.decode(metadata, { ...context, ism });
-
-      case IsmType.ROUTING:
-        return DefaultFallbackRoutingMetadataBuilder.decode(metadata, {
-          ...context,
-          ism,
-        });
-
-      case IsmType.ARB_L2_TO_L1:
-        return ArbL2ToL1MetadataBuilder.decode(metadata, {
-          ...context,
-          ism,
-        });
-
-      default:
-        throw new Error(`Unsupported ISM type: ${ism.type}`);
     }
   }
 }
