+++ typescript/sdk/src/ism/types.ts
@@ -15,18 +15,9 @@ import {
 import type { Address, Domain, ValueOf } from '@hyperlane-xyz/utils';
 
 import { OwnableConfig } from '../deploy/types.js';
+import { ZHash } from '../metadata/customZodTypes.js';
+import { OwnableSchema, PausableSchema } from '../schemas.js';
 import { ChainMap } from '../types.js';
-
-import {
-  ArbL2ToL1IsmConfigSchema,
-  IsmConfigSchema,
-  MultisigIsmConfigSchema,
-  OpStackIsmConfigSchema,
-  PausableIsmConfigSchema,
-  TestIsmConfigSchema,
-  TrustedRelayerIsmConfigSchema,
-  WeightedMultisigIsmConfigSchema,
-} from './schemas.js';
 
 // this enum should match the IInterchainSecurityModule.sol enum
 // meant for the relayer
@@ -167,3 +158,100 @@ export type RoutingIsmDelta = {
   owner?: Address; // is the owner different
   mailbox?: Address; // is the mailbox different (only for fallback routing)
 };
+
+const ValidatorInfoSchema = z.object({
+  signingAddress: ZHash,
+  weight: z.number(),
+});
+
+export const TestIsmConfigSchema = z.object({
+  type: z.literal(IsmType.TEST_ISM),
+});
+
+export const MultisigConfigSchema = z.object({
+  validators: z.array(ZHash),
+  threshold: z.number(),
+});
+
+export const WeightedMultisigConfigSchema = z.object({
+  validators: z.array(ValidatorInfoSchema),
+  thresholdWeight: z.number(),
+});
+
+export const TrustedRelayerIsmConfigSchema = z.object({
+  type: z.literal(IsmType.TRUSTED_RELAYER),
+  relayer: z.string(),
+});
+
+export const OpStackIsmConfigSchema = z.object({
+  type: z.literal(IsmType.OP_STACK),
+  origin: z.string(),
+  nativeBridge: z.string(),
+});
+
+export const ArbL2ToL1IsmConfigSchema = z.object({
+  type: z.literal(IsmType.ARB_L2_TO_L1),
+  bridge: z.string(),
+});
+
+export const PausableIsmConfigSchema = PausableSchema.and(
+  z.object({
+    type: z.literal(IsmType.PAUSABLE),
+  }),
+);
+
+export const MultisigIsmConfigSchema = MultisigConfigSchema.and(
+  z.object({
+    type: z.union([
+      z.literal(IsmType.MERKLE_ROOT_MULTISIG),
+      z.literal(IsmType.MESSAGE_ID_MULTISIG),
+      z.literal(IsmType.STORAGE_MERKLE_ROOT_MULTISIG),
+      z.literal(IsmType.STORAGE_MESSAGE_ID_MULTISIG),
+    ]),
+  }),
+);
+
+export const WeightedMultisigIsmConfigSchema = WeightedMultisigConfigSchema.and(
+  z.object({
+    type: z.union([
+      z.literal(IsmType.WEIGHTED_MERKLE_ROOT_MULTISIG),
+      z.literal(IsmType.WEIGHTED_MESSAGE_ID_MULTISIG),
+    ]),
+  }),
+);
+
+export const RoutingIsmConfigSchema: z.ZodSchema<RoutingIsmConfig> = z.lazy(
+  () =>
+    OwnableSchema.extend({
+      type: z.union([
+        z.literal(IsmType.ROUTING),
+        z.literal(IsmType.FALLBACK_ROUTING),
+      ]),
+      domains: z.record(IsmConfigSchema),
+    }),
+);
+
+export const AggregationIsmConfigSchema: z.ZodSchema<AggregationIsmConfig> = z
+  .lazy(() =>
+    z.object({
+      type: z.literal(IsmType.AGGREGATION),
+      modules: z.array(IsmConfigSchema),
+      threshold: z.number(),
+    }),
+  )
+  .refine((data) => data.threshold <= data.modules.length, {
+    message: 'Threshold must be less than or equal to the number of modules',
+  });
+
+export const IsmConfigSchema = z.union([
+  ZHash,
+  TestIsmConfigSchema,
+  OpStackIsmConfigSchema,
+  PausableIsmConfigSchema,
+  TrustedRelayerIsmConfigSchema,
+  MultisigIsmConfigSchema,
+  WeightedMultisigIsmConfigSchema,
+  RoutingIsmConfigSchema,
+  AggregationIsmConfigSchema,
+  ArbL2ToL1IsmConfigSchema,
+]);
