+++ typescript/sdk/src/router/types.ts
@@ -11,15 +11,11 @@ import { Address, AddressBytes32 } from '@hyperlane-xyz/utils';
 import { HyperlaneFactories } from '../contracts/types.js';
 import { UpgradeConfig } from '../deploy/proxy.js';
 import { CheckerViolation } from '../deploy/types.js';
+import { HookConfigSchema } from '../hook/types.js';
+import { IsmConfigSchema } from '../ism/types.js';
+import { ZHash } from '../metadata/customZodTypes.js';
+import { DeployedOwnableSchema, OwnableSchema } from '../schemas.js';
 import { ChainMap } from '../types.js';
-
-import {
-  DestinationGasSchema,
-  GasRouterConfigSchema,
-  MailboxClientConfigSchema,
-  RemoteRoutersSchema,
-  RouterConfigSchema,
-} from './schemas.js';
 
 export type RouterAddress = {
   router: Address;
@@ -68,3 +64,40 @@ export interface RouterViolation extends CheckerViolation {
 
 export type RemoteRouters = z.infer<typeof RemoteRoutersSchema>;
 export type DestinationGas = z.infer<typeof DestinationGasSchema>;
+
+export const MailboxClientConfigSchema = OwnableSchema.extend({
+  mailbox: ZHash,
+  hook: HookConfigSchema.optional(),
+  interchainSecurityModule: IsmConfigSchema.optional(),
+});
+
+export const ForeignDeploymentConfigSchema = z.object({
+  foreignDeployment: z.string().optional(),
+});
+
+const RemoteRouterDomain = z.string();
+const RemoteRouterRouter = z.string().startsWith('0x');
+export const RemoteRoutersSchema = z.record(
+  RemoteRouterDomain,
+  RemoteRouterRouter,
+);
+
+export const RouterConfigSchema = MailboxClientConfigSchema.merge(
+  ForeignDeploymentConfigSchema,
+).merge(
+  z.object({
+    remoteRouters: RemoteRoutersSchema.optional(),
+    proxyAdmin: DeployedOwnableSchema.optional(),
+  }),
+);
+
+const DestinationGasDomain = z.string();
+const DestinationGasAmount = z.string(); // This must be a string type to match Ether's type
+export const DestinationGasSchema = z.record(
+  DestinationGasDomain,
+  DestinationGasAmount,
+);
+export const GasRouterConfigSchema = RouterConfigSchema.extend({
+  gas: z.number().optional(),
+  destinationGas: DestinationGasSchema.optional(),
+});
