+++ typescript/utils/package.json
@@ -17,6 +17,7 @@
     "@types/sinon": "^17.0.1",
     "@types/sinon-chai": "^3.2.12",
     "chai": "4.5.0",
+    "eslint": "^8.57.0",
     "mocha": "^10.2.0",
     "prettier": "^2.8.8",
     "sinon": "^13.0.2",
@@ -36,6 +37,7 @@
     "build": "tsc",
     "clean": "rm -rf ./dist",
     "check": "tsc --noEmit",
+    "lint": "eslint src --ext .ts",
     "prettier": "prettier --write ./src",
     "test": "mocha --config .mocharc.json './src/**/*.test.ts'",
     "test:ci": "yarn test"
